resume screening/requirements.txt
Flask==2.3.2
nltk==3.8.1
PyPDF2==3.0.0
pymongo==4.5.0

sentence-transformers==2.2.2
scikit-learn==1.3.0
google-generativeai
